fix:PostAPi

+++ GniApi/Dtos/RequestDto/LoanRequest.cs
@@ -4,13 +4,13 @@ namespace GniApi.Dtos.RequestDto
 {
     public class Loanrequest
     {
-        public int amount { get; set; }
+        public decimal amount { get; set; }
         public int productCode { get; set; }
         [JsonIgnore]
         public bool limitExceeded { get; set; }
         public int percentageRate { get; set; }
         public int comissionPercent { get; set; }
-        public int repaymentAmount { get; set; }
+        public decimal repaymentAmount { get; set; }
         public int repaymentPeriod { get; set; }
     }
 }
